Breytingar
Er buinn að bæta inn nokkrum layerum:
-presentation: seem geyser user interface
-domain: sem sækjir og sendir
- data: tenging í .txt skjöl
