--- Listi.cpp
@@ -1,15 +1,0 @@
-#include "Listi.h"
-
-List::List(string proName, string proGender, int proBirthday, int proAge)
-{
-    string name= proName;
-	string gender= proGender;
-	int birthdate= proBirthday;
-	int age= proAge;
-}
-
-void getname()
-{
-	return name;
-}
-
