+++ include/data.h
@@ -1,0 +1,13 @@
+#ifndef DATA_H
+#define DATA_H
+
+
+class data
+{
+    public:
+        data();
+    protected:
+    private:
+};
+
+#endif // DATA_H
