+++ include/domain.h
@@ -1,0 +1,13 @@
+#ifndef DOMAIN_H
+#define DOMAIN_H
+
+
+class domain
+{
+    public:
+        domain();
+    protected:
+    private:
+};
+
+#endif // DOMAIN_H
