+++ src/data.cpp
@@ -1,0 +1,6 @@
+#include "data.h"
+
+data::data()
+{
+    //ctor
+}
