+++ src/domain.cpp
@@ -1,0 +1,6 @@
+#include "domain.h"
+
+domain::domain()
+{
+    //ctor
+}
