+++ src/presintation.cpp
@@ -1,0 +1,16 @@
+#include "presintation.h"
+
+
+presintation::presintation(string proName, string proGender, int proBirthday, int proAge)
+{
+    string name= proName;
+    string gender= proGender;
+    int birthdate= proBirthday;
+    int age= proAge;
+}
+
+void getname()
+{
+    return name;
+}
+
